Move hell effect messages into the miscellaneous database.
This way, they can be much more easily expanded.  There are a few
deficiencies, however, compared to the monster speech database, which it
seems inappropriate for since monsters are not speaking exclusively.

1. Embedded lua isn't run, so the "You smell brimstone./Brimstone rains
from above." message, dependent on you.can_smell(), is currently reduced
to the second message in all cases.

2. There's no way to change what channel the hell messages come through,
although the relevant code from mons_speaks_msg() could probably be
adapted.  Also, if (1) is fixed, there is a lua equivalent of noisy()
that could be used for the sound channel messages, although according to
its surrounding comments, it's limited compared to the non-lua version.