Tweak unbreathing mutation description.
Unbreathing refers to the ability to survive without breathing, no more,
no less.  Being able to go underwater is still a separate quality, and
being able to speak presumably still requires the ability to breathe, so
it still applies to e.g. Zin-worshipping grey draconians.

As for mummies (not to mention felids and octopodes), assume they speak
via magical means, as I can't think of a different justification right
now.