Tweak spirit fade summons, mark them as finished.
They now summon either an angel or a cherub, or rarely an apis, phoenix,
daeva, or pearl dragon, as before. Instead of being guaranteed to get
hit with a nasty, though, you could just get a pip. The weights may need
to be tweaked further to make them less annoying.