Change always_confirm_butcher to allow butchering the top corpse of a stack with one keypress
"auto" is the default and identical to the current default behaviour -
if there's a single corpse, butcher it, if there are multiples, ask
which ones to butcher.

"always" is identical to the old "always_confirm_butcher = true" - the
prompt is always displayed, even if there's only one corpse.

"never" is new behaviour - if there's a single corpse, it's butchered,
and if there are multiples, then the top corpse is butchered (skipping
corpses that your god would get angry at you for butchering entirely).
Possibly this should be expanded to also skip rotten corpses if you're
not a saprovore, and so on.