Add more pan lord part tiles from Denzi (#4888).
I also changed rim 1 to rim 0 - some of the heads are meant to blend into the bodies.

I also removed demon_head_heads - the concept is nice, but it didn't look great on any of the combos I saw while testing, and looked bad on many. demon_head_tentacles looks bad on some of the combos as well, but great on others, so I'd rather keep it.