Changes to minmay_elemental_castle to make it actually generate sometimes.
Changed to ORIENT: southwest instead of float, with a small adjustment to
the layout to make this work. The vault was simply too big to work as a
floating vault and would never get placed (at least, I wasn't able to get it
to work after about 50 tries).

Even with this change it's not very likely to show up.