Make scrolls of silence more common.
They had the pointlessly low weight of 10/10000 (note that this is less
common than acquirement). I've increased the chance of you finding these to
the same as acquirement, EWIII and summoning. Thus, still uncommon, but
you'll at least find a few now. The weight of detect curse has been lowered slightly as well.