Use the level stack to save return target for portal vaults/Abyss/Pan/Zigs/Lab.
Unlike the initial implementation that was talked about, this one nearly
completely ignores what stair is used to do X, meaning you can exit a branch
by any stair that goes up, etc.

Also, this commit earns the prize of the most rebased and edited one.  And it
ended up a tiny fraction of its initial size, with far simpler code to boot.