Make spider_temple_joshua a normal vault (not a minivault)
This is a fairly large map and I can't see a reason for it to be a
minivault. I've given the map ORIENT: north and moved it to large.des.

There are some other huge minivaults, but they aren't as problematic as this
one when it comes to placement (this one would almost always fail when I
tested it).