Don't use "you feel drained" as part of Boris' speech.
Boris has a couple of "fake spells", one of which can cause the "You feel
drained" message to appear. This is bad because it can cause mummies,
ghouls, etc to "feel drained" even when that should be impossible. I changed
the message to "you feel disoriented".

I don't like how boris has fake spells like this in the first place, but if
he's going to have them, they should at least be semi-plausible.