Make Zin's protection from mutations better than rMut.
The amulet has a 10% chance of letting a mutation through.
Zin now has 5% at 160 piety, 2% at 180, immunity at 200 (like before).

The protection working reliably only just near 200 was ok when Zin's piety
didn't decay and sinks sucked.  Nowadays, giving people an incentive to
micromanage piety to pickup that pile of gold when a neqoxec comes is not
fun.

As with any buff, we may want to nerf Zin in some other way.