Make Deep Dwarf Artificers always use rods; make monster Striking do damage.
50% of artificers would just be generic melee dwarves. Also the chance of
them having rare rods was only 1/100. Now it's 1/25. If this is too high
it's not a big deal, since you won't find these guys outside of zigs anyway.

This means that most of them will have a rod of striking. I increased the
damage for monster striking (was 1d8), since it was low enough that it might
as well have been 0. I'm pretty sure they'd do more damage throwing darts.

This introduces an asymmetry between player striking use and monster
striking use (the rod is complete garbage for players). If this is a problem
then I think striking should be buffed on the player end. Note that players
can't start with one any more.