Don't refer to the missing expired portal tile for Spider.
There's actually one draw existing on Mantis, but I don't even bother
putting it in since the portal vault will be removed before 0.10 (ie, soon).
It's unlikely we can balance the _branch_ in time, but having the portal and
branch exist together causes confusion in the code and extra work, so the
former is already gone in the level_type overhaul.

Actually, we might just go and remove the portal outright.