Add some missing commits from UCC merge
During my period of inactivity (June 21 to October) master received a
handful of commits that appear to have never made it into UCC and are
now missing. These appear to be isolated to fight.cc and do not seem
to follow a pattern (the commits missing do not seem to follow entirely
in sequence, for example).

In any case, this adds in all that I can find of those missing commits.