Tweak nooodl_heptagram.
Made the central area less open, reduced the number of monsters in the
side rooms, and chopped off some unneeded bits. The game can still spawn
huge numbers of monsters though, if the rng wills it.

One thing about this vault is that it still has no corridors to fight in,
which might be a bad thing for such a large vault.