Make monster_info::pos grid- instead of player-relative.
It was converted back everywhere it was used anywhere, and it became
useless after the player moved. Mostly, the monster position is needed
to refer back to the grid cell the monster is in, so as long as
map_knowledge uses the normal grid coordinate system, monster_info
should too.

Note that Webtiles obscures coordinates in a better way already
anyway, by using the coordinate of the first cell sent as origin and
sticking to that coordinate system as long as the level doesn't
change.