Save the game immediately on entering wizmode, to prevent cheating.
If you manage to crash the game (which is sadly easy in wizmode), you can
use this to roll back to the last saved state.  This hardly ever matters as
typically if you can enter wizmode you could mess with the save directly,
but let's fix it just in case a public server enables wizmode for untrusted
users.