Name the wizard (with custom spells) in the laboratory_2 minivault.
In general, I think that vault monsters should be given custom
names/descriptions if they have custom stats/spells. In this case, the vault
monster has the polymorph other spell.

I am calling it a "mad wizard" here because it is basically a mad scientist
theme. No other information is given to the player, but this should at least
clue them in that it's not a normal wizard monster.

Note that the monster did already have a custom colour defined. However,
this isn't enough because (a) custom colour is a lot less obvious than a
custom name, and (b) it looks the same in tiles.