Use a more() when Evolution gives you a mutation.
Since it can come at any moment, it's likely you'd miss it.

This version is hard coded, it'd be better to have this configurable, but
for an easy way it'd require coming up with a message given AFTER the
mutation, and I can't think of a good one right now.