Modifications to lemuel_temple_firewalk vaults.
Both of these vaults guarantee a safe corridor to the temple entrance...
the problem was that the clouds could still appear close enough to this
corridor for a travel exclusion to be placed over it. This could have the
effect of "blocking" the temple entrance with exclusions, making it likely
for the player to miss the entrance to the temple (and misleading them into
thinking there's no safe area).

I moved the cloud generators a bit to guarantee that the middle corridor
will never have an exclusion placed on it.

I also added a floor tile to the "marked" variant.