Add more cloud data to map_knowledge; make tilepick.cc use it instead of the actual cloud_struct.
This adds the cloud tile and the value of cloud.decay/20, clipped to
the range 0-3, to the map_knowledge, and organizes all cloud-related
data into a new cloud_info struct. Both of these are required for the
tile picking code.

(Doesn't introduce a new minor tag, so saves from the previous commit
will crash.)