Fix antimagic() issues.
First, it was previously possible to die instantly by reading a scroll of
vulnerability while levitating over water or lava... it looks like this was
an unintentional consequence of the delay changes in 0.6. Now you should get
a turn to do something before the levitation ends.

Second, reading a scroll of vulnerability with teleport active used to
function as an instant teleport... a very powerful trick that isn't really
in the spirit of antimagic. Now it just cancels the teleport. This is a buff
to monsters with antimagic breath (quicksilver dragons and purple dracs), but
that seems okay... quicksilver dragons have been getting fewer kills with
each version for some reason, and no kills so far in 0.10.