Change seen_context from strings to enums, slightly untangle it.
It's still a terrible mess.  I didn't manage to understand it well enough
so all[1] functionality is preserved, at the cost of clarity.  I unified
some values that are functionally identical, but not yet all.

[1]. The context is reported only as a number in crash dumps, a minor
regression but 1. saves speed, 2. it'd be a matter of copying the enum
names if someone actually needs this.