Adjust the weights for some Elf:5 vaults.
There are a lot of Elf endings now. Most have normal weight (10) but there
were a few oddities: a few had weight 3 (this gave them about a 1/20 chance of
appearing in a game) and the classic one had a weight of 50 (with another
old classic having an effective weight of 20).

I can understand the rationale behind having the classic maps being a bit
more common, but I think 50 is too high, so I've reduced it to 30. This will
cause it to be picked in about 1/5 of all games. It might be worth just
reducing it to 10 like the other maps; I'm not sure if "being classic" is an
appropriate justification for it being more common than the others.