Move Dwarf entries to the start, like in any other branch .des file.
It made sense in blue_anna's initial concept, when there was a 5-level
branch with a separate portal vault ("Dwarven Tomb") enterable from the
bottom of it.  He didn't get to making the actual portal so the entries
are repurposed as entries to Dwarf rather that exits from Dwarf to that
Dwarven Tomb.