Fix item sacrifices using the unidentified value for unidentified items.
This affects Elyvilon, Nemelex, and Jiyva. With Nemelex, we should be on the
watch for whether weapons and armour tend to have too high a weight now for
deck selection... they already tended to get very high weights if you
sacrificed everything, and this makes it worse by increasing the value of
enchanted (or ego) items that you sacrificed without identifying.