Replace some translucent stone with translucent rock/grates in vaults.
This is on account of the recent nerf to smite-targetted spells. This made
it so some vaults had monsters which were unkillable without (real) smiting
or LRD/shatter. Now you can use digging to get to them.

Grates are used in cases where the vault still uses undiggable (stone) glass in
other parts, in order to clearly mark the difference. Otherwise regular
diggable (rock) glass is used.