Apply the smite restrictions to monsters as well.
Such a massive change just makes no sense to be one one way.  Removal of
true smite from players was done for balance purposes (can be exploited with
some vaults or with vitrification) rather than due to theme, but we'd drown
under bug reports if we introduce such an inconsistency.