Remove "skill exp" from PlaceInfo and char dumps.
It had not been useful since the 0.9 skill system overhaul, and often
had negative values because training happened before calculating the
gain.  This removes the field entirely, as well as the associated
column "F" of the kills_by_place table.  The old column "G" is now
calculated based on level XP only.