Auto-id evil weapons for Elyvilon worshippers.
A problem: if you don't intend to worship evil, you can switch to Ely, and
get a free global weapon id.  Blocking id until *** piety conflicts with the
recent max_piety change (which needs to be rethought, though), and would be
mostly an interface hassle anyway -- so a better idea is welcome.