Show menu for =a by default if auto_list is on.
Also, replace msg::stream << with mpr/mprf.

Note that you (still) cannot use the menu for the second prompt (letter
to change to).  This would be desirable, but is a somewhat more
complicated because, unlike list_spells, choose_ability_menu returns an
index rather than a hotkey.

Furthermore, the message ("No such ability.") may be wrong, but we have
no way of telling whether the menu was exited because of a wrong
character or because of pressing escape.

Fixes #4074, #5019.