Fix several vaults broken by recent smite targetting changes.
By "several" I mean at least 20. I might have missed a few!

Vaults which rely on monsters being able to attack through glass now use
iron grates (which certain spells can go through, for both players and
monsters). For the most part this doesn't change much, but there is the
issue of grates being diggable. This changes the behaviour of a few vaults
(eg david_glass_crypt).

I've added comments to the ones affected by this (hopefully it is only a
temporary solution). In the long term, it might be good to add an undiggable
"steel grate" dungeon feature, depending on whether the changes to
targetting stay as they currently are.

Another minor issue is that some of these vaults look weird in tiles now.

Also, a couple of "zoo" entry vaults which used grates now use normal glass, to
block smite-abuse.