Don't create disconnected floor when ruining levels.
When we decide to ruin a wall or door, instead of replacing it with
floor, choose a random adjacent (non-wall, non-door) feature.  Thus,
when a wall borders only impassable features (trees, statues, deep
water, ...), it will be replaced with an impassable feature, preserving
connectivity.

This could be kind of silly if we "ruin" a wall into a statue, but that
can be fixed later.