Even more careful ruination.
Never replace a door with an impassable feature: use either floor
or water.  Never replace a feature with a statue, since that's
silly (in fact, let statues help support the adjacent walls).

Trees, water, floor, and lava are the features that can still replace
a ruined wall; and shallow water and floor for a ruined door.