Make unrand melee effects happen a bit later.
It was being called during handle_phase_attempted, before calculating
damage; this meant that Devastator was doing nothing.

The new location is just after player_monattk_hit_effects.  This is
a little later than pre-UCC (it had been in the middle of that
function), but this should be safe, and allows monsters and players
to use the same call.