Warn when using elec brand in water near allies.
This should prevent accidentally offending gods.  The prompt is
given a little too often: the attack is actually safe if the defender or
the ally is flying or is known to have rElec, but we check only whether
the cell contains water.

As part of this implementation, the smite targetter can now take a
predicate to select only certain cells to be affected.