Prompt about chopping drainable, not just bottleable, corpses
Since can_bottle_blood_from_corpse() checks whether the player has the
bottle ability, it was always returning false for low-level vampires.
Instead, we should give the confirmation prompt if the player could
*either* drain the corpse or bottle it.

The prompt is given for poisonous corpses when the player has rPois.
However, we do not check for whether the player is capable of eating
at all; perhaps we should.