Disambiguate the LOS_SOLID confusion.
Its name suggests it's about line of effect, while in fact it's both effect
and sight.  Thus, I've split its uses into LOS_SOLID and LOS_SOLID_SEE -- the
former is what LOS_SOLID was usually understood to mean, the latter is what it
actually was (being targettable with an arrow/beam).

The difference is: LOS_SOLID_SEE obeys clouds.

The name of the latter is abysmal -- if you have a better idea, please sed.