Use a formatted_scroller for the mutations screen (edlothiol).
A Menu does not allow multiple colours per entry, so we instead use
a formatted_scroller in display_mutations (for non-vampires; vampires
are still treated differently).

Make the '*' marking form-based mutations yellow, and the '+' marking
non-physical vampire mutations lightred.

Also, do not include the legend and "Press '!'..." in chardumps; and
avoid duplicating the draconian AC message.