Mark scales as partially suppressed in statue form.
This necessitates distinguishing partially-active from hunger-based
activity, so add MUTACT_HUNGER and use that for non-physical mutations
on vampires between satiated and very full.

Also add a function to properly format fake mutations that depend on
form.

Note that, for these purposes, "thin skeletal structure" and "distortion
field" do not count as scales (they are fully active in all forms).