Be more careful about marking mutation status.
For one, don't mark a mutation as partially suppressed when your hunger
level is high enough to support its current level (e.g. a level one
mutation for a Satiated vampire).

Secondly, do not mark the mutation status when asking about a specific
mutation level rather than the player's mutation level (level == -1).
This keeps the (())*+ annotations out of the notes list and the output
of the wizmode &] command.