Avoid doubled articles on features, and fix descriptions.
If the feature's name begins with the word "some", "a", "an", or "the"
(or uppercase versions), do not add an article.

Add "the" to the name of "open sea", and "endless lava", since those
do not make sense with "a".  Adjust dat/descript/features.txt
accordingly.

Also adjust dat/descript/features.txt to restore the description
text for floor (useless though it may be).