Remove a "magic number" from unarmed attack delay.
Previously unarmed combat reached min delay at 25 skill, meaning that the
benefits of raising UC skill drop off sharply at that point. This change moves
min delay all the way to 27 skill (while keeping min and max delay constant,
so this is a slight UC nerf).

It would of course be good to remove the "magic numbers" for reaching min
delay with weapons as well, but that would be a far larger change.

The formula for bat attack delay was also changed.