Reduce frequency of cloud traps.
Each one was given a chance of 40%, making you likely to get several per
level. The comment in the des file, as well as the commit message for these
traps, indicated that this weight was extra-high for testing purposes.

There aren't any serious bugs left with these and there has been a lot of
gameplay feedback for them. Unfortunately, there was no hint as to what the
"normal" chance for these traps is supposed to be, so I've reduced it to a
chance of 10% for now (still likely to see several per game, but not several
per level).

I have also disabled two problematic cloud traps which have almost no effect
on gameplay.