Nerf the starting ambush in evilmike_geh.
The starting room no longer has a chance to spawn random enemies (eg
tormentors, hellions) which aren't placed specifically by the vault. Also,
the four fire giants start just outside of your line of sight, instead of
all of them being in view.

This is intended to make the room easier to survive without needing to
resort to stair dancing.