Avoid spurious regeneration of tiledef headers.
tilegen had a bug causing it to think that small (< 1024 byte) files had
changed when they had not.  This resulted in tiledef-dngn.h (and thus
the source files that depended on that) being regenerated even when
nothing had changed.  Whether the bug triggers depends on uninitialised
memory contents, so might or might not have shown up in a given run.