Revert "Rename a few artefacts to accomodate the "the " prefix."
It made them sound abysmal, and breaks the very reason for "Lehudib's
crystal spear" to exist.

What you wanted to fix might be slightly awkward, but it was considered to
be good enough for years, and thus is certainly better than reversed names.
(0.9 used "the cursed -1 Lear's chain mail", this commit restores that).

This reverts commit 003f65b5a3bf00f9eb827b457359a19dcac4c590.
This reverts commit e262dd72296f6bc4f2ace00ccb77b7e160c94da8.