Veto the holy pan level if you are with a good god.
This is a cheap solution, but it fixes what I think is a very serious
problem with the vault (and to a large extent, the very concept of a holy
pan level).

Basically, if you are with a good god and have reasonable piety, the level a
total non-threat. Any loot in the level is free loot, and this is bad thing:
TSO and Zin do not need buffs in the extended endgame. Even without loot,
the vault would just be a harmless "setpiece" with 0 actual gameplay. So,
for now, the vault simply will not appear if you're with a good god. Let's
say it's because the lords of Pandemonium will only allow you to enter the
holy level if you're likely to kill everything there.

This commit also tweaks the loot a bit in the level.