Fix regression with enslaved souls and the MF_FIGHTER flag.
The MF_FIGHTER flag is the equivalent of M_FIGHTER for individual
monsters, and is applied for enslaved souls where the original monster
has M_FIGHTER, e.g. a soul-enslaved orc warrior will be a spectral orc
with this flag set.  The melee code now properly checks for this again,
instead of checking only for M_FIGHTER on the base type.