Increase, never reduce, rPois for undead.
This is modeled after the change in 91711a7.  The special case that
necessitated its always being 1 for undead was the split between demonic
and undead abominations, in which the former had one level of poison
resistance, and this split is no longer in the code.

There are only two odd cases now regarding poison resistance for undead:
vampire mosquitoes and profane servitors.  Due to the previous code, the
former's poison vulnerability was effectively replaced with poison
resistance, and the latter's poison resistance was doubled despite being
a copy of angels' poison resistance.  Both have only one level of poison
resistance now, by virtue of being undead and not due to flags.