Fix save breakage from the constriction merge.
The new minor version tag was inserted *before* an existing tag, causing
saves with the previous minor tag to fail to load.  This swaps the tags
to the correct order.

This will break saves made from the constriction branch prior to its
merge, but only a few developers would have such saves.

Fixes #5105.