Change the Gehenna floor colour from yellow to brown.
This is in line with the similar change to Tomb recently. Using a yellow
floor makes it hard to see where TSO halos extend to - slightly problematic
in a branch that many people use TSO for.

I've elected to use brown here because it's fairly close to yellow, and is a
better choice than red (the walls are already red).