A couple of changes to hell ending colours/tiles.
Geh:7 and Coc:7 both give special descriptions for the floor, which weren't
working. This fixes that.

In addition to the custom floor descriptions, custom colours were given to
the floor. This commit extends that to the floor tiles now. They look different
from the rest of the branch, but I don't consider this an inconsistency if
it goes to the trouble of giving it a custom description. Plus, these are
full-level vaults, so they can get away with looking a bit different.

Finally, a few cosmetic changes have been made to evilmike_geh.