Change hints on gaining Spellcasting 1.
Now that spellcasting skill isn't required for memorising spells, the
"You're starting to get the hang of this magic thing." message is no
longer needed.  Furthermore, change the hints-mode hint to reflect this,
and to mention some of the benefits of Spellcasting skill.

Also, give the hit upon gaining the first level of spellcasting, even if
another magic skill is higher.