Build databases in parallel.
They are disk seek/fsync() bound, so it's a gain even on 1-way systems.
The gain is way smaller than I expected yet non-negligible; the worst case
is Windows playing over SMB which could sometimes be as bad as 115 seconds
with the server on the same idle Ethernet segment.

This improves only db not des; the latter is fsync free though.