Stop constricting when not adjacent.
This is still done a little late (only when the constrictor tries
constricting again, not as soon as the actors are non-adjacent).

There is no message for this case.  If one of the actors moves before
constriction damage is attempted, there is the normal escape attempt,
and the resulting message.  This should also be cleaned up.