Eliminate duplicated symbols.
Objects, even const ones, should not be defined in header files, or
there will be one copy for each source file that included that header.
Move the offending definitions into source files, and replace them
with extern declarations in the headers.

Also, make the opc_* instances const and not static.