Fix #5133: secret door positions revealed next to walls with redefined tiles.
This was a regression from my tilepick.cc changes. Since the tile
picking code now operates on map_knowledge, it didn't know about the
undetected secret door and thus didn't use the tile override from
adjacent tiles.

This is a bit of a problem with tile_flv, which on the one hand is
player-accessible (parts of it get transmitted to the webtiles client,
for example), but on the other hand could contain "secret" data, like
in this case (maybe the only one at the moment). (Note that this
"secret" data didn't actually get sent to the client, but it could
easily have been if more tile picking logic was on the client side.)

I've opted to fix up tile_flv directly in tile_init_flavour, so that
secret doors take the tile override of their neighbours. Of course,
that causes the problem that the tile override is still there when the
secret door gets turned into an open door, which resulted in an
assert ("non-door tile").

That's why now, only certain tiles are allowed as door overrides; all
others simply do nothing for doors since they are assumed have been
set for undetected secret doors. It also means that you can either
override the appearance of an undetected secret door, or its
appearance after detection, but not both. I don't think that was done
anywhere, but if it was, a more complicated solution might be
necessary.