Do not lose constriction when teleporting.
Add a flag (default true) to actor::move_to_pos(), actor::moveto(), and
move_player_to_grid() indication whether clear_far_constrictions()
should be called.  When teleporting or blinking, pass false as that
flag, pull along constrictors/constrictees, and only then call
clear_far_constrictions().

Eronarn points out that it would be better to use an enum for the type
of motion, rather than the (now two) boolean flags.  I agree, but this
is a much bigger change so I have not done so yet.

It is still possible to get "comes along for the ride x2" if the
teleporter is contricting and constricted by the same actor.