Remove -cTele from six baileys and an ice cave.
Level flags like no_tele_control should be used sparingly, and there wasn't
any flavor-reason why a random subset of the baileys and an ice cave should
have it. cTele is useful for stealing loot in almost any situation, and this
should be addressed on a more general level, not on a vault-by-vault basis.

Making controllable teleports a finite resource by making cTele not work on
teleportitis or a ring of teleportation would help with this. Note also that
cTele targetting changed in 0.9 so that you often need mapping to use it
effectively, and mapping is usually limited as well.

I left no_tele_control in volcano_caves since cTele avoids the main point of
the map, but I think there is probably a better way of doing this... for
instance, the loot and monsters in blocked-off tunnels could be destroyed if
the player is not present in them when the tunnel collapses.