Abyssal punctuated equilibrium.
Morphing now happens in "jerks".  During a jerk, the morphing rate is
about three times what it used to be; between jerks, it is nearly zero.
Jerks are around 100--300 turns apart, depending on abyss speed
(200--600 for Cheibriadites), and last for about 1/3 of that time.

For some reason, loading old saves causes some major jerking for a few
turns before reaching the slow initial speed.  I'm not clear why
this is happening.