Make abyss somewhat more open, much less liquid.
Previously a "wall" had a 1/2 chance of being a wall (1/6 each for rock,
stone, and metal) and a 1/2 chance of being a liquid (1/6 each for lava,
deep water, and shallow water).  Change this to 1/3 rock, 1/6 stone,
1/12 metal, 1/12 green crystal, 1/12 lava, 1/12 deep water, 1/6 shallow
water.  Make fewer "walls" to compensate for the decreased openness.

When an old save is loaded, features will change on the next tick
(though obstacles will remain obstacles and floors floors).  There
doesn't seem to be much that can be done about that.