Take away the beak from monster octopodes.
Note that the difficulty of these monsters is not set in stone; they don't
occur anywhere in places they're anything more than a gimmick so if anyone
has a need to change it, anything from kobolds to draconians or even liches
is ok.  Monster octopodes' main purpose is merely show_player_species anyway.