Prevent zotdef summon crashes.
This was supposed to be fixed by aacbd52, which prevented the check
from firing if the monster is on a no_rtele_into tile (such as the
loot squares).  However, there were a few problems:

 1. The zotvault5 loot vaults had three floor cells that were not
    no_rtele_into.  Turn them into permaglass instead.

 2. Zotvault1 and zotvault2 had some loot vaults that were surrounded by
    rock and possibly stone, with one permaglass window).  If you dig
    the rock wall, the resulting floor tiles would not be no_rtele_into,
    so summoning could then cause a crash.  Surround those vaults with
    permarock.

The second issue might not be likely to happen accidentally, but could
be used to cause intentional crashes and "time travelling": if things
are going badly, crash the game and restore the last save.

Fixes: #4830.