Actually seed the libc RNG.
Due to an oversight it wasn't being seeded at startup unless -seed
was provided.  It's not used for very much, but it is used for
std::random_shuffle() and, in particular, the god list in
initialise_temples().  As a result, the same gods would get overflow
altars each game (except for the second and subsequent games under
restart_after_game).