Count zotdef games as 10 times as long for scoring purposes.
It was far to easy to hit 100 megapoints.  The average winning score was
in fact 83 megapoints, and would be even higher were it not for the cap.

As kilobyte points out, scoring isn't really relevant for winners, since
the amount of XP you get is mostly based on luck, and rune timing is
fixed.