Scale constriction damage with time.
Use the time scaling code which was used for clouds.
Should fix #5168, at least for player constriction. I'm not sure if it's
right for monsters. It doesn't check the actual energy used, just the base
monster speed.
It should bring back anaconda to a more reasonable threat level. It does
raise slightly the damage for nagas.