Tweak generation of unique Pandemonium lord levels.
Pandemonium tended to "front load" itself with unique levels, with the first
one coming up very early in your visit, and the last one taking quite a
while to show up (usually about 9 levels after the previous one).

This changes it so (a) the chance of a unique level starts low and increases
over time (rather than decreasing), and (b) there's less randomness overall:
an absurdly unlucky player (1 in 1000) might take 65 levels rather than
100. You're likely to start seeing uniques about 10 levels in now. Overall it
still takes an average of 27 levels for all four, so the length of your visit
should be about the same as it was before.

The reason for this change is to give Pandemonium a better sense of
progression and pacing. Unique levels are the closest thing Pan has to
"branch endings", so if anything they should become more common the deeper
you get. Throwing the player into the deep end and then forcing them to go
through 10 easy levels doesn't make much sense.