Tweak milestones and notes for death/banishment/pacification/enslavement.
Previously milestones for pacified monsters happened when the pacified
monster disappeared (i.e. by leaving the level), not when it was
actually pacified. This is now fixed, successfully using Yred's Enslave Soul
on a monster is recorded as well, and all four of these ways of defeating
a monster now give different milestones and notes.