Don't block 'a'bilities from having Perfect success rates.
They were previously capped at the high end of Excellent, for no apparent
reason. A couple of abilities were allowed to reach Perfect (mostly ones
that start that way). The blink mutation was also special cased to be
able to reach this level, along with Nemelex's Draw One.

As for the rest, they got capped at a 99.9% success rate, which is high
enough that it has almost no tactical effect. The main effect is obscuring
things for the player, by not letting them know when their ability success
rate has maxed out.