Fix #5111: Items under a plant aren't indicated [in tiles].
The problem was that tile picking happened concurrently to the show
update, i.e. while the map cell was being built. This meant that when
placing the plant tile, the item wasn't yet in map_knowledge, so no
indicator was added.

I've refactored the code so that all tile picking happens at the end
of show_update_at. I fear this may lead to new bugs, but the
alternative (going back to directly accessing igrd) would have been
hacky and would likely have caused new (if small) information leaks.