Fix #5201 (the ?-mark for multiple items in tiles had gone missing).
This was another victim of the tile picking refactoring. The problem
was that tile_place_item tested if there were more items by checking
item.link, which doesn't work for an item from a map_cell.

While I was at it, I also removed a few outdated comments and made the
tile_place_* functions static.