Fix #5204: Unknown wands and potions on ground show as flame wands and potions of curing.
The function get_item_info sets the sub_type to 0 if it is unknown (or
rather, leaves it at 0). The problem is that that's a valid item type,
and the tile picking code had no way of knowing for sure if the item
simply had sub_type 0 or was unidentified.

I opted to always add the KNOW_TYPE flag to the item_info when the
type is known. This seems like the cleanest solution and also helps
simplify the tile picking code (which doesn't need to check
you.type_ids anymore).