Replace success adjectives with failure rate percentages. (dtsund)
This is slightly modified from the patch submitted here:
https://crawl.develz.org/mantis/view.php?id=5179

The original patch uses fractions instead of percentages, but I found this
to be quite visually cluttered and unhelpful.

After much fretting about the best way to implement this, I settled on
rounding up to 1% if your failure rate is below 1% but above 0. I tried some
alternatives, but I find this way to be the easiest for me to visually parse
at a quick glance at my spell list. If this is unideal, it's very easy to
change.