Blame inner flame damage and clouds on the inner flame's caster.
It was hard-coded to the player, which 1. caused arena to assert-crash, 2.
gave the player penance in the "experimental stone golem" vault.

It may be a good idea to be consistent with spores and give blame (and xp)
to those who set off agentless inner flames, but currently there's no way
to tell apart those cast by a dead hostile caster from those created by
a vault.