Give Deep Dwarf Artificers (the monster) wands instead of rods of striking.
They can still rarely get a random rod, but most will have wands.
Usually low-tier attack wands, since this monster fills the role of a weak
ranged attacker, and can show up in a fairly early vault.

This also reverts the change that introduced an asymmetry with rods of
striking. It's back to 1d8 now for players and monsters. It's probably worth
just removing rods of striking at this point.