Colour failure rates based on how likely the spell is to cause a dangerous miscast effect.
In the casting and memorization list, "12%" is now coloured one of four colours
based on the chance that casting the spell will cause a severity 2 or 3
miscast. This does not take into account Sif or Kiku's protection or penance.

The cutoff points are currently at 0.001, 0.005, and 0.025, and the colours
used are lightgrey, yellow, lightred, and red.