Various changes to vault-defined feature descriptions.
This removes usage of portal_desc to rename stone walls (they will now
appear as regular stone walls). This works around a problem where, when
destroying the wall, the floor would keep the description.

This also makes the elephant statues in due_jungle_book indestructible, to
work around the same problem (note that this is already done in other
vaults). It would be good to have a better solution to this in the long term.